sem3_mcp_task/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
fastmcp>=0.2.0
kafka-python>=2.0.2
pyspark>=3.5.0
clickhouse-driver>=0.2.6
avro>=1.11.3
py4j>=0.10.9.7
requests>=2.28.0
mcp>=0.1.0
fastavro>=1.9.0
prometheus-client>=0.17.0
starlette>=0.27.0